backend/requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
itsdangerous==0.24
pymongo==2.7.2
tornado==4.0.2
wsgiref==0.1.2
oauthlib==0.7.2
python-twitter==2.2
requests==2.5.3
requests-oauthlib==0.4.2
